requirements.txt
discord.py[voice]==2.3.2
yt-dlp==2024.8.6
PyNaCl==1.5.0